Synchronizationv2

Closes #10

See merge request myahye/cmpt431-IronIV!5

+++ DAC/src/command/Command.java
@@ -3,8 +3,8 @@ package command;
 import java.io.Serializable;
 
 public interface Command extends Serializable {
-	public int getTimeStamp();
-	public void setTimeStamp(int timeStamp);
+	public long getTimeStamp();
+	public void setTimeStamp(long timeStamp);
 	public int getConnectionID();
 	public void setConnectionID(int connectionID);
 }
